A few Dynmap changes.
Currently when someone says something while viewing the website map, the chat only stays up for 5 seconds, then disappears (this includes your own text).  I suggest bumping it up to 15 seconds.  

The second change is to allow us to scroll back in text.  The way we have it now, once the messaagettl expires, the chat is just cleared out.  It would be nice to be able to scroll through the text so you don't have to be glued to the map. The default is 100 lines, which I would think would be more than enough.